updated player lives - player struct to volatile and changed it so that only one missile can hit a player per game loop.

+++ source/same70/SAME70_SpaceInvaders/SAME70_SpaceInvaders/src/Game/sprite.h
@@ -68,7 +68,7 @@ typedef enum
 //player
 typedef struct
 {
-	uint8_t numLives;
+	int numLives;
 	uint32_t x;
 	uint32_t y;
 	uint32_t sizeX;
